implementado relatório da movimentação de estoque por dia. alterada estrutura e layout das views.

+++ app/Http/Controllers/ProductController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Product;
+use App\Models\PurchaseOrder;
+use App\Models\SalesOrder;
+use Illuminate\Support\Facades\DB;
 
 class ProductController extends Controller
 {
@@ -14,7 +17,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::all();
+        $products = Product::orderBy('name')->get();
 
         return view('admin.product.index')->with('products', $products);
     }
@@ -97,8 +100,51 @@ class ProductController extends Controller
     {
         $product = Product::find($id); 
 
+        $purchase_orders_relations = PurchaseOrder::where('product_id', $id)->count();
+        $sales_orders_relations = SalesOrder::where('product_id', $id)->count();
+
+        if(($purchase_orders_relations + $sales_orders_relations) > 0){
+            return redirect()->back()->withErrors(['Cannot delete due existing relations with purchase or sales orders.']);
+        }
+
         $product->delete();
 
         return redirect()->route('product.index');
+    }
+
+    public function report(){
+        $purchase_orders_query = DB::table('purchase_orders')
+                                    ->where('purchase_orders.amount','>', 0)
+                                    ->select(
+                                        DB::raw(
+                                                'product_id, 
+                                                products.name, 
+                                                products.sku, 
+                                                purchase_orders.amount, 
+                                                DATE_FORMAT(purchase_orders.created_at,"%Y-%m-%d") as created_at, 
+                                                \'purchase\' as type'
+                                                )
+                                            )
+                                    ->join('products','product_id','=','products.id');
+
+        $orders = DB::table('sales_orders')
+                            ->where('sales_orders.amount','>', 0)
+                            ->select(
+                                DB::raw(
+                                        'product_id, 
+                                        products.name, 
+                                        products.sku, 
+                                        sales_orders.amount, 
+                                        DATE_FORMAT(sales_orders.created_at,"%Y-%m-%d") as created_at,
+                                        \'sale\' as type'
+                                        )
+                                    )
+                            ->union($purchase_orders_query)
+                            ->join('products','product_id','=','products.id')
+                            ->orderBy('created_at', 'DESC')
+                            ->orderBy('name', 'ASC')
+                            ->get();       
+
+        return view('admin.product.report')->with('orders', $orders);
     }
 }
